+++ AuthPermissions/SetupCode/BulkLoadRolesService.cs
@@ -3,14 +3,13 @@
 
 using System;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using AuthPermissions.DataLayer.Classes;
 using AuthPermissions.DataLayer.EfCode;
 using AuthPermissions.PermissionsCode.Internal;
-using AuthPermissions.SetupParts.Internal;
+using AuthPermissions.SetupCode.Internal;
 using StatusGeneric;
 
-namespace AuthPermissions.SetupParts
+namespace AuthPermissions.SetupCode
 {
     public class BulkLoadRolesService
     {