+++ AuthPermissions/SetupCode/DefineUserWithRolesTenant.cs
@@ -6,7 +6,7 @@ using System.ComponentModel.DataAnnotations;
 using AuthPermissions.DataLayer.Classes;
 using AuthPermissions.DataLayer.Classes.SupportTypes;
 
-namespace AuthPermissions.SetupParts
+namespace AuthPermissions.SetupCode
 {
     public class DefineUserWithRolesTenant
     {
@@ -17,17 +17,17 @@ namespace AuthPermissions.SetupParts
         /// <param name="roleNamesCommaDelimited">A string containing a comma delimited set of auth roles that the user</param>
         /// <param name="userId"></param>
         /// <param name="uniqueUserName">A string that is unique for each user, e.g. email. If not provided then uses the userName</param>
-        /// <param name="tenantName">Optional: The unique name of your multi-tenant that this user is linked to</param>
+        /// <param name="tenantNameForDataKey">Optional: The unique name of your multi-tenant that this user is linked to</param>
         public DefineUserWithRolesTenant(string userName, string roleNamesCommaDelimited,
             string userId = null,
-            string uniqueUserName = null, string tenantName = null)
+            string uniqueUserName = null, string tenantNameForDataKey = null)
         {
             UserId = userId; //Can be null
             UserName = userName ?? throw new ArgumentNullException(nameof(userName));
             RoleNamesCommaDelimited = roleNamesCommaDelimited ??
                                       throw new ArgumentNullException(nameof(roleNamesCommaDelimited));
             UniqueUserName = uniqueUserName ?? UserName;
-            TenantName = tenantName;
+            TenantNameForDataKey = tenantNameForDataKey;
         }
 
         /// <summary>
@@ -47,8 +47,14 @@ namespace AuthPermissions.SetupParts
         /// </summary>
         public string UniqueUserName { get; private set; }
 
-        public string TenantName { get; private set; }
+        /// <summary>
+        /// This contains the Tenant name. Used to provide the user with a multi-tenant data key
+        /// </summary>
+        public string TenantNameForDataKey { get; private set; }
 
+        /// <summary>
+        /// List of role names in a comma delimited list
+        /// </summary>
         public string RoleNamesCommaDelimited { get; private set; }
     }
 }