+++ AuthPermissions/SetupCode/TenantTypes.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) 2021 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT license. See License.txt in the project root for license information.
 
-namespace AuthPermissions.TenantParts
+namespace AuthPermissions.SetupCode
 {
     /// <summary>
     /// This defines the types of tenant the AuthPermissions can handle